Lots of work and improvements

+++ core/include/giygas/PushConstantsRange.hpp
@@ -1,11 +1,12 @@
 #pragma once
+#include <cstdint>
 
 namespace giygas {
 
     class PushConstantsRange {
     public:
-        size_t size;
-        size_t offset;
+        uint32_t size;
+        uint32_t offset;
     };
 
 }
