+++ core/include/giygas/RenderPass.hpp
@@ -13,7 +13,7 @@ namespace giygas {
 
     class RenderPassCreateParameters {
     public:
-        size_t attachment_count;
+        uint32_t attachment_count;
         const RenderPassAttachment *attachments;
     };
 
@@ -23,7 +23,7 @@ namespace giygas {
         virtual RendererType renderer_type() const = 0;
         virtual void create(const RenderPassCreateParameters &params) = 0;
         virtual bool is_valid() const = 0;
-        virtual size_t attachment_count() const = 0;
+        virtual uint32_t attachment_count() const = 0;
         virtual const AttachmentPurpose *attachment_purposes() const = 0;
     };
