+++ core/include/giygas/UniformBuffer.hpp
@@ -9,7 +9,7 @@ namespace giygas {
     public:
         virtual ~UniformBuffer() = default;
         virtual RendererType renderer_type() const = 0;
-        virtual void set_data(size_t offset, const uint8_t *data, size_t size) = 0;
+        virtual void set_data(uint32_t offset, const uint8_t *data, uint32_t size) = 0;
     };
 
 }
